fix: org-roam-ql children, descendents, and first-children index predicates
Previously, the implementation that checks whether one index numbering is a
child or descendant of another index numbering would use string-prefix-p. This
overlooks the cases when, say, index1 is "1" and index2 is "19": it would return
non-nil incorrectly.

This fixes these cases by checking the parts of the index numberings directly.